feat: Add SSHClient signal to RunnerCommand
This commit adds the `sshClientSignal` to the `RunnerCommand` class in order to track the SSH client used for connecting to the server. This signal is set to `null` by default and is updated with the SSH client instance when connecting to the server. This change allows other methods in the `RunnerCommand` class, such as `setCompleted`, to access the SSH client and perform necessary operations.