feat: Retry job search after 1 second if no jobs found
This commit adds a retry mechanism to the `startJobSearch` method in the `RunnerCommand` class. If no jobs are found and the progress signal is not null, the progress signal will be updated to indicate that no jobs were found and the retry will happen after 1 second instead of the fixed 10 seconds. This change improves the responsiveness of the job search functionality.